requirements.txt
beautifulsoup4==4.2.0
requests==1.2.2
sinaweibopy==1.1.2
